+++ src/components/AppRoutes.jsx
@@ -6,6 +6,7 @@ import {
 
 import SolutionsView from './builder/solutions/SolutionsView';
 import SolutionDetailsView from './builder/solutions/SolutionDetailsView';
+import CreateSolutionview from './builder/solutions/CreateSolutionView';
 import ArchitectureView from './builder/ArchitectureView';
 import BillofMaterialsView from './builder/bom/BillofMaterials';
 import ServiceDetailsView from './builder/services/ServiceDetailsView';
@@ -71,8 +72,9 @@ class AppRoutes extends React.Component {
         return (
             <Routes>
                 <Route path='/' element={<LandingPage user={this.props.user} addNotification={this.props.addNotification} />} />
-                <Route path='/solutions/user' exact element={<SolutionsView user={this.props.user} addNotification={this.props.addNotification} isUser />} />
                 <Route path='/solutions' exact element={<SolutionsView user={this.props.user} addNotification={this.props.addNotification} />} />
+                <Route path='/solutions/user' exact element={<SolutionsView user={this.props.user} addNotification={this.props.addNotification} isUser />} />
+                <Route path='/solutions/new' exact element={<CreateSolutionview user={this.props.user} addNotification={this.props.addNotification} />} />
                 <Route path={'/solutions/:id'} element={<SolutionDetails user={this.props.user} addNotification={this.props.addNotification}/>} />
                 <Route path='/boms/user' exact element={<ArchitectureView user={this.props.user} addNotification={this.props.addNotification} isUser/>} />
                 <Route path='/boms/infrastructure' exact element={<ArchitectureView user={this.props.user} addNotification={this.props.addNotification} isInfra/>} />
